--- packages/shared/src/components/search/SearchBarInput.tsx
@@ -1,21 +1,0 @@
-import { FormEvent, InputHTMLAttributes } from 'react';
-import { SearchChunk } from '../../graphql/search';
-import { SearchBarSuggestionListProps } from './SearchBarSuggestionList';
-
-interface SearchBarClassName {
-  container?: string;
-  field?: string;
-  form?: string;
-}
-
-export interface SearchBarInputProps {
-  valueChanged?: (value: string) => void;
-  showIcon?: boolean;
-  showProgress?: boolean;
-  className?: SearchBarClassName;
-  onSubmit?: (event: FormEvent, input: string) => void;
-  inputProps?: InputHTMLAttributes<HTMLInputElement>;
-  chunk?: SearchChunk;
-  shouldShowPopup?: boolean;
-  suggestionsProps: SearchBarSuggestionListProps;
-}
